Add test result references
For each test added svg files with layers needed for checking if
resulting layout is correct. These files can be generated by running
tests with `--save-results-as-reference`. After these files are
generated, they should be manually validated by human. If layout is ok,
it can be saved and then used as reference for automatic tests.