Use KiCad's `CONNECTIVITY_DATA` for better collision detection
This should fix some edge cases which may occur when using custom diode
position, in particular situation where first track segment coming out
of diode pad is so short that the next one connecting to it would be
marked as colliding with that pad (due to clearance). With connectivity
data in use it is easier to ignore such collision (although the
`TestTrackCollision` method blew out of proportion).