Stop using silkscreen layer to test examples result
Ignore silkscreen svg when asserting results. Plugin does not do anything on those layers
and maintaining them is a bit tedious, for example between 7.0.0 and 7.0.5 there is very
slight difference in silkscreen digits which would be falsely detected as failure here.
`generate_render` will still produce silkscreen svg to have nice images in test report but
for checking result it is ignored.