Add raw KLE layout parsing
Introduce `kle_serial.py` which is able to convert RAW layout file to
its internal form. This will be an built-in python alternative to [1]
which currently needs to be used outside this plugin (for example via [2]).
When integrated, plugin will be able to automatically convert RAW layout
if provided by user. Now it supports only internal layouts files.

Reference files for testing are generated by patched kle-serial [3].

[1] https://github.com/ijprest/kle-serial
[2] https://adamws.github.io/kle-serial/
[3] https://github.com/adamws/kle-serial ee49e0e8f44a975b3ace4dc81d3997e420e1fa61