Fix `TestTrackEndpointDangling` error when running with KiCad 7.0.7
Needed to rework kicad version checking because checking only MAJOR
version is not sufficient. Normally should not use tuples for comparing
versions but this is best we can do without dependencies (usage of
`packaging` package would be recommended). We assume that version is
always three numbers only.

Added required argument to function call for KiCad 7.0.7

Fixes #8