Add support for proper GUI teardown without process kill
This would be required for upcoming coverage support.
Process kill worked just fine but it did not allow for collecting
(saving) coverage measurement. In case dialog is started by pytest,
add input listener which must be used for closing the application.
When started normally, it should behave just like before.