Allow placement of switch elements without layout file
Move extra elements placing to independent loop, which runs even
when layout is not provided. This loop uses already placed switches
as reference points.

This change hits some rounding difference hence on kicad6
some reference files needed to be updated.
This is not a problem when comparing with kicad7 references
because those use less significant digits for float values.

Double checked and looks like some rotated diodes
position are different on insignificant precision digit so it can be
safely ignored and not considered as an error with code.
For 2x3-rotations example, D3 before and after changes from this
commit:
Before: `Footprint: D3 position: (73969323, 44670728)`
After:  `Footprint: D3 position: (73969322, 44670728)`
(that's 1 nanometre difference)