Start testing with more complex switch footprints
Added reversible Kailh hot-swappable footprint and added two tests for
automatic diode routing. This footprint has multiple pads with the same
number so `FindPadByNumber` can't be used in order to set netlist for
all of them.
Added tests check if diode is connected with single track to nearest pad
of the same netlist.