Improve track templates handling for switch-diode routing
This is major improvement of user drawn connection template between
first switch and a diode pair which is then replicated for remaining
pars.
Now it supports switch footprints which have more than one
pad of the given netcode (this is the case for reversible switch
footprints). It also supports vias due to usage of different KiCad API
than before.